Users can read GET /system
In certain cases (for certain tokentypes) users need to be able to
read the system config of these tokens.
So we added another staticmethod to the tokenclass.
A token may define, which token settings are saved "public"ly.
A user may read these public settings:
like the questions of the questionnaire token.

Also reworked the before and after methods.
Put those in a central file, thus all other API files
have no before, after and error handler anymore.

Closes #289
Closes #287
