+++ privacyidea/lib/decorators.py
@@ -115,3 +115,4 @@ def check_copy_serials(func):
         return f_result
 
     return check_serial_wrapper
+
